Unify mistral and mixtral
